Copied project from Monogame
Just fixed all type errors and try to start. Most of the stuff actually
works. Boxes are small and not seen. Need to work on Camera
and resizing on windows.

But a very good start. Also seems to have waaaaaay more performance than
MonoGame. But first need to fix everything before comparing it.