Added Dic2 and Removed State
# Added Dic2

Added the Dic2 data-structure that basically just is a Dictionary that
contains another dictionary for keys. In Perl would just be
$hash->{$key}{$entity} = $value.

Reason for this addition is performance. Previously i had two dictionaries
for View. One containing the Visible and another the Hidden View. But
it was somehow complex to insert something into one the other. Put
the idea how i would it make in Perl. There i would just do

$hash->{true}{$entity} = $whatever

I rebuild this two dimensional selection with Dic2. Technically it could
get deeper levels, but at least in .Net i can create compund keys. For
example i wanted to have the combination of Visible and the Layer.

In Perl could be.

$hash->{$visible}{$layer}{$entity} = $data

But in .Net at least i can make a compund key out of visible and Layer.
In F# just a tuple (visble,layer).

Now instead of creating an array in Systems.draw and let it sort by
the layer i just iterate through (visible,BG3), then (visible,BG2)
and so on.

No Array creation and sorting needed anymore. Framerates jumped from
~1100 fps up to ~1400+ fps.

# Removed State<>

This wasn't really needed anymore. At some time before State was more
complex, it added an Event whenever a entity was added/removed so
a Cache could be generated. But i removed and simplified this
some time ago. Was basically just a wrapper around a Dictionary,
now removed it completely.

Some code needed to change.