Updated Benchmark Numbers
Maximum fps for lower box counts is lower. But for many boxes framerate
are higher. Makes sense as Dic2 has a little bit more overhead. Getting
a View is now always fetched with 2 Dictionary lookups.

But with many objects the creation and sorting of array is now gone
and this saves up more performance as just 2 dictionary lookups. So
overall a performance improvement when many many boxes/enemys are drawn.