Added basic object culling
Object Culling by just checking if position of whatever should be
rendered is inside the visibility of the camera + some offset.

At the moment the values are hardwired from my virtual Resolution.
Maybe will be configurable later. Otherwise must be adjusted.

Had no performance impact when everything is still drawn, even when
i expected it. As it has an additional if-check in the drawing loop.
Just improved fps greatly.