Started refactoring of Components
I change the API to be like Raylib, one namespace with all functions
inside is better than having 10 modules with some functions.

The problem with the modules was also to find the one you really wanted.
For example there can be `View.create...` functions, but sometimes you wanna
convert from one type to another. So you come up with `View.from...`
function naming.

But sometimes this is not convenient to have for later defined modules.
So we also end up with `Sheet.toView` because a `View.fromSheet` was
not possible at that point. This clutter now ends. One module and all
functions that create something start with `create<Type><Extra>`

On top i think this excessive mapping common in F# to never use record
syntax and map everything to functions is also annoying. So a record
with 6 fields ends up with a single create constructor and 6 "getter"
function and 6 "setter" functions, even if no single one is needed.

Mapping to a function without arguments is more a pain to work with.
Having a function call with 3+ arguments where no argument is somehow
named is just a guessing game, hard to read or understand.

And the setter have the problem that they create a new structure on each
invocation. A waste of cpu/memory and heap allocations.

So people tend to write.

let newThing =
    Module.empty
    |> Module.withA foo
    |> Module.withB bar
    |> Module.withC baz

seriously, what is wrong with

let newThing = {
    Module.empty with
    A = foo
    B = bar
    C = baz
}

It also does not need to create two intermediate records that are
immediately thrown away. Its even less code to type. As F#8 also now
supports nested changing this is in my opinion anyway a prefered way to
write and use records.