Changed Transform
For Future improvement I changed the way how Transform are built. Previously
i had a transform, and every Transform had an optional Parent. This
allowed me to position things relatively to a parent. When a parent
moves, rotates and scales (never testes it) it applies its to
its children.

However it has some disadvantages. Because i never knew the globalPosition
of an entity without calculating it. Also when i render everything
i had to calculate the final Transform and then depending on it
show something.

But having the real global position is crucial for some further improvement.
For example when i want to build a QuadTree i need the global Position
of every object. Not just the local position.

So i changed everything To either be Local or a Parent Transform.

This change makes accessing position, rotation and so one a little bit
annoying, but i provided some members and functions that makes it
less annoying.

now rendering changes. Instead of iterating through all entities,
calculating the global position and then issue a Draw command i do
the processing and updating of the global position in its own step.

Now rendering just can render everything without ever calculating
the positions. I expected that framerates drop, because the Transform
position is more complex, basically its now a pointer of a pointer
instead of a single pointer.

But overall it improved performance. Okay performance without defining
a parent stays the same. But there is basically no performance drop
anymore when i assign a parent to an object.

On top of all of it, now computating of the global position can be
put on its own thread, something i couldn't do before. Having access
to the global positions and calculating them now gives me more options
to implement further optimizations.

Also because both Transforms are now divided, i can built two dictionaries
where on Dictionary only contains LocalTransforms and another dictionary
that only contains Parent Transforms.

In the Transform update i now can only update those Transform that have
a parent, instead of iterating through all Transforms. In a game with
lot of mixed Transforms this can greatly improve and reduce calculations.

Only those that needed to be processed needs to be processed.

So overall impressed by the change. Especially as i expected that fps
goes initially down, but instead it improved further. Seems calculating
parent and issuing DrawCalls to the GPU had massive CPU cache misses
or something like that. So just seperating those makes stuff faster.