Added Storage Back Again
Already implemented a Storage data-structure but deleted it as it had no
advantage. But now i added it back again. But the implementation
is different than before.

Before i used a Dictionary and an Array and did some Dictionary mapping
to save the current entry in an array. This made insert and remove O(1).
Even indexing was O(1) as i directly could fetch an entry.

Theoretically this DS is worse, but at some point is far easier implemented
than the previous DS. But sometimes the easier code can still perform
better. Now i use a Sorted Array, indexing is binary search
and inserting is Insertion Sort. Removing is basically insertion sort
reversed.

Even if this DS could perform worse than a Dictionary, i will probably
stick with it, because it allows me to directly access all data as an
array that allows me todo easy Multi-Threading.