Storage is Crap
Seriously why do i even try this again and again. I guess i have written
like 10 benchmarks and all kind of crap to always prove again and again
how worse an Array with Binary Search is compared to a Hash.

This shows it again. Rendering 10,000 fps before was 650+ fps, now it
droped down to 60 fps. Its a performance drop by 10x.

I am not even sure that this magnitude of a performance drop can be
outweighted by any kind of multi-threading. Maybe i switch back to my previous
Storage. It had less of a performance decrease as this piece of crap.