Back to old Storage implementation
The old Storage implementation i had before was implemented with an additional
dictionary. Instead of inserting key/values with an insertion sort
and doing binary search it just saves a key -> index mapping in an
additional hash.

Insertion just can add new elements at the end. Its still O(1) but
a little bit more costly then just using a dictionary alone.

Removing is also O(1) as the element to be removed is just swaped with the
last element. For the swapped element the KeyToIndex just needs to be
updated. Maybe is a little bit confusing but performans to
a normal dictionary basically without any performance penalty.

Why all of this complicated stuff?

Because it saves all values just in an ResizeArray that i have acces to
and can directly iterate, and even process it in Parallel.