Added Multi-Threading
Added Multi-Threading back again, and now make it the default. But
this didn't change fps much as somehow would expect. This has maybe
multiple reasons.

Even with 90,000 boxes shown what can be multi-threaded is updating
a vector (Movement component) and Calculating the Global Position
with the Parent. But this in usual doesn't take much time. Even with
90,000 elements those computation are usually not much for modern CPUs

Improvement with multi-threading are basically only seen with very high
boxes count like 90,000. And then they are still very minor.

But overall I optimized the system so far that assign a parent doesn't
seems to have any kind of computational overhead anymore to the
unparented versions.

At least when gameobject only have a parent depth of 1.

Performance doesn't go up primarily because Rendering or calls to
DrawTexture() will likely eat up 90%+ of all the CPU timing.

But overall this is still good. because in the future i can add more and
more component now, even with some computational overhead and can put
all kinds of computation to other threads. Some systems also can be
run in parallel.

So main thread just starts all kind of threaded computation, waits for
everything to finish and just needs to render. This system will scale
probably nicely in the future.

Maybe I can even further optimize the Storage data-structure and directly
implement a hash myself instead of building it together. But at the moment
this is simply not needed.