Changed Storage implementation
Changed Storage implementation. Before i had two Dictionaries
for key->index and index->key mappings, was a little bit confusing
to understand, changed it in the way that now every key,value pair
is directly stored in the array.

In the playground i had and doing micro-benchmarking it actually shows
that this approach is actually a little bit slower.

I repeatedly iterate through everything 50000 times and measure the average
of one iteration. with two dictionaries it takes 0.3ms to iterate once.
with the struct approach it is around 0.45ms. Not much of a difference.

But using in a real scenario like here it makes no difference at all what
makes somehow sense, as it is just that 0.1ms less or more of a single frame
it takes.

But benchmarking in scripting environment has not all optimizations enabled.
And somehow i believe that using a struct maybe can be better in those
cases I iterate through an array and anyway need the key and the
value.

So i stick with this approach. Also a big benefit is that the code is
easier. When performance is literaly the same i just stick to the code
that is easier to understand.