This didn't worked as expected
I assumed that a System.Random object is created once, and then used
across multiple calls. But Profiling resulted that whenever random()
is called a new System.Random object was created. This is not
only horrible in performance but actually wrong as this way does not
produce good random values. Just deleted the whole functions on the
Extension. Don't need them and i create random direction in another
way in the future.