Merged Transform back to a Single State
I made this "optimization" because System.Tranform.update then
only need to update those elements that has a Transform applied. But this
leads to an overhead getting the transform, as i have to try two Storages
to test instead of a single one.

This optimization before made the Transform.update maybe better, but caused
more problems in Systems.View.draw. But the first can and is parrelised
and the View.draw cannot be parrellized at all. So now made it
that View.draw runs better.

Also now changed the test-becnhmark so half of all moving boxes are
parent and the other is not.