Changed Movement
Made 3 Components out of it instead of one. Did this change to primarily
make it mutable. Before it was completely immutable and the data cannot
be changed without creating new object and inserting.

But don't know if the change is actually bad. Seems to have more stutter.
Also activated SetTargetFPS() to limit to a certain fps todo some testing.