Disable manual GC
First. It should only collect 0. 0 is short-lived object and 1 to 2
are long lived objects. GC.Collect(2) always created a full foreground GC
that blocks.

Optimized says that it doesn't need to immediately run, and the false
issues a background GC. Actually more what i wanted.

But deactivated it at the moment as i realized that there happens a lot
of GCs already. It seems all of it comes from creating strings. Even
in an empty scene it creates around 8 MiB of garbage per second.

Deactivating some DrawText command immediately reduced it to 4 MiB.

So yeah nice. because strings are objects and created on the heap they
create the greatest garbage in the system. But at least in a final game
this should be no problem anymore as there i would textures most of
the time. Not that it creates problem currently.

Still works, but not what i expected. I will look if i can avoid it and
create strings on the stack.