Explicitly re-add GC.Collect() calls
doing it explicit still seems to trigger GC more often, what is a good
thing because then less garbage is collected and pause times are less.

I encountered that with 90,000 boxes gc pause can go up 12ms sometimes.
Calling it explicitly reduces it to maybe 6ms when it runs.

With just 40,000 boxes it usually is <1 ms.