Moved fixedUpdate handling to main function
Also changed it into a while loop so fixedupdate could be potential
set to a higher value than fps, so it truly becomes frame-rate indepent.

But this also has a problem when a computer cannot keep up anymore when
somehow the CPU the game runs is too slow. But i can live what that.
Just buy a better PC.