Changed Transform
After some Profiling and looking into compiled code with ILSpy i found
out that the previous Transform with Local|Parent created code that
used object casting with `is` to either cast to Local or Parent.

In Profiling this showed up as the the highest cost in updating the
Transform. 3/4 of the time was just spent casting obj to either Local
or Parent. 1/4 time was spent in getting the Transform from the Storage.

As Transform is a core data-structure and pretty much will probably the
most used data-type i think this is inacceptable. So changed the code back
to just a single Transform.

Profiling and code generatin actually showed a lot better performance.
For example instead of 3s spent in System.Transform.update it now reduced
to 0.5s spent in that function.

In extreme situation like 90,000 boxes fps jumped from 60~fps to 80~fps.

So this change gave a good performance boost.

But still i wonder why the F# compiler does this stuff. The F# compiler
usually generates a Tag field so every object can be directly casts with
`as` instead of doing `is` casts. But maybe those make no difference?

Well, anyway this way it is just faster. It also made some code a little
bit easier, so i guess it is an overall good change.