Use Task Library directly
I am suspicious to the exact way how the task builder is used. In Tests
and some tests it showed that all tasks finished always in the exact same
order in the way they are specified. But when different Task really run
in Parallel there should not always be in the exact same order.

Using TPL directly like now actually produces this result. So i stick
with this approach.