Updated Benchmarks
Actually they didn't really changed. But this time i waited maybe a
minute before noting the fps. The fps starts a litte bit higher but
the longer the program runs it goes down a little bit. This reason
for this should be that i constantly add new objects and delete
from AutoMovement and AutoTarget. This creates some memory fragementation
and reduces fps over longer running periods.